requirements.txt
Flask>=0.10.1
-e git+git://github.com/mjdesrosiers/dota2py#egg=dota2py
-e git+git://github.com/smiley/steamapi#egg=steamapi
-e git+git://github.com/mjdesrosiers/pypubg#egg=pypubg
requests>=2.4.3
pymongo>=2.8
parsedatetime>=1.5
python-dateutil>=1.5
nltk>=3.0.5
azure-cognitiveservices-search-imagesearch>=1.0